requirements.txt
pillow
numpy
opencv-python
reportlab
requests
nbformat
canvas